view usage page modified

+++ lib/presentation/page_component/package_info_card.dart
@@ -156,7 +156,7 @@ class PackageInfoCard extends StatelessWidget {
 }
 
 class _DataUsageBar extends StatelessWidget {
-  final int usedGB;
+  final double usedGB;
   final int totalGB;
 
   const _DataUsageBar({required this.usedGB, required this.totalGB});
